Reverting separating components
